Improve v1/v2 separation in tests
