+++ tests/execute_command.py
@@ -1,0 +1,14 @@
+import asyncio
+
+
+async def execute_command(cmd, **kwargs):
+    proc = await asyncio.create_subprocess_shell(
+        cmd,
+        stdout=asyncio.subprocess.PIPE,
+        stderr=asyncio.subprocess.PIPE,
+        **kwargs,
+    )
+    stdout, stderr = await proc.communicate()
+    if proc.returncode != 0:
+        raise RuntimeError(stderr.decode("UTF-8"))
+    return stdout.decode("UTF-8").strip()
