+++ tests/v1/04_api/test_task_collection_api.py
@@ -26,7 +26,7 @@ from fractal_server.tasks.v1._TaskCollectPip import _TaskCollectPip
 from fractal_server.tasks.v1.background_operations import (
     background_collect_pip,
 )
-from tests.fixtures_tasks_v1 import execute_command
+from tests.execute_command import execute_command
 
 PREFIX = "/api/v1/task"
