+++ tests/v2/03_api/test_api_task_collection.py
@@ -12,7 +12,7 @@ from fractal_server.syringe import Inject
 from fractal_server.tasks.utils import get_collection_path
 from fractal_server.tasks.utils import get_log_path
 from fractal_server.tasks.v2._TaskCollectPip import _TaskCollectPip
-from tests.fixtures_tasks_v1 import execute_command
+from tests.execute_command import execute_command
 
 
 PREFIX = "api/v2/task"
